Version 1.1
Fixed bug that limits the download below 100 images, also added option to add sorting arguments.

+++ Image_extractor.py
@@ -2,33 +2,39 @@ import urllib.request
 import urllib.error
 import json
 import os
+import time
 import config
 
-def genRedditUrl(subreddit, sort, postLimit):
-    if sort != "":
-        sort = "/" + sort
-    url = 'https://www.reddit.com/r/' + subreddit + sort + '/.json?limit=' + str(postLimit)
+downloadCount = 0
+errorCount = 0
+
+def genRedditUrl(subreddit, sortType, sortArg, after):
+    if sortType != "":
+        sortType = "/" + sortType
+    if sortArg != "":
+        sortArg = "&" + sortArg
+    url = "https://www.reddit.com/r/" + subreddit + sortType + "/.json?limit=100" + sortArg + "&after=" + after
     return url
 
 def getJson(url):
-    header={'User-agent': 'Image-extractor 1.0'}
+    header={"User-agent": "Image-extractor 1.1"}
     req = urllib.request.Request(url=url, headers=header)
     try:
         response = urllib.request.urlopen(req)
     except urllib.error.HTTPError as e:
-        print('HTTPError: '+str(e.code))
+        print("HTTPError: "+str(e.code) + "\n")
     except urllib.error.URLError as e:
-        print('URLError: '+str(e.reason))
+        print("URLError: "+str(e.reason) + "\n")
     else:
         data = json.loads(response.read().decode("utf-8"))
         return data
 
 def extractImageUrl(jsonFile):
     ret = {}
-    posts = jsonFile['data']['children']
+    posts = jsonFile["data"]["children"]
     for element in posts:
-        url = element['data']['url']
-        title = element['data']['title']
+        url = element["data"]["url"]
+        title = element["data"]["title"]
         asciistr = title.encode("ascii", errors="ignore").decode()
         unclean = str(asciistr)
         cleantitle = unclean.translate(str.maketrans(" ","_","/.\\?#:;*<>\"'|"))
@@ -37,36 +43,52 @@ def extractImageUrl(jsonFile):
     return ret
 
 def downloadImg(urls,category):
+    global downloadCount
+    global errorCount
     saveDir = mkSaveDir(category)
     for key, value in urls.items():
         extentionLocation = value.rfind(".")
         fileName = (saveDir + key + value[extentionLocation: extentionLocation + 4]).encode("utf-8")
-        if not os.path.exists(fileName):
-            print("Saving: " + fileName.decode("utf-8") + '\n')
-            urllib.request.urlretrieve(value, fileName)
-
-
+        if (not os.path.exists(fileName)) and (downloadCount < config.downLimit):
+            downloadCount += 1
+            try:
+                urllib.request.urlretrieve(value, fileName)
+                print("Saving img no." + str(downloadCount)+ ": " + fileName.decode("utf-8") + "\n")
+            except urllib.error.HTTPError as e:
+                print("Img no." + str(downloadCount) + " HTTPError: "+ str(e.code) + "\n")
+                errorCount += 1
+            except urllib.error.URLError as e:
+                print("Img no." + str(downloadCount) + " URLError: "+ str(e.reason) + "\n")
+                errorCount += 1
+            
 def mkSaveDir(dir):
     if not os.path.exists(dir):
         os.makedirs(dir)
-    return dir + '/'
+    return dir + "/"
 
 def runDownloader():
     for item in config.subreddit:
-        print("--------------------------------------------------" + '\n')
-        print("Starting downloads for: " + item + '\n')
-        print("--------------------------------------------------" + '\n')
-        url = genRedditUrl(item, config.sort ,config.postLimit)
-        jsonFile = getJson(url)
-        imgDict = extractImageUrl(jsonFile)
-        downloadImg(imgDict,item)
-        print("--------------------------------------------------" + '\n')
-        print("Done downloading: " + item+ '\n')
-        print("--------------------------------------------------" + '\n')
+        global downloadCount
+        global errorCount
+        print("--------------------------------------------------")
+        print("Starting downloads for: " + item)
+        print("--------------------------------------------------" + "\n")
+        after = ""
+        while downloadCount < config.downLimit and after != None:
+            time.sleep(3)
+            url = genRedditUrl(item, config.sortType, config.sortArg, after)
+            jsonFile = getJson(url)
+            imgDict = extractImageUrl(jsonFile)
+            downloadImg(imgDict,item)
+            after = jsonFile["data"]["after"]
+        print("--------------------------------------------------")
+        print("Done downloading " + item + " Error Count: " + str(errorCount))
+        print("--------------------------------------------------" + "\n")
+        downloadCount = 0
+        errorCount = 0
 
 def main():
     runDownloader()
-
 
 if __name__ == "__main__":
     main()
